[ADD] warehouse loading
