+++ lib/widgets/warehouse/warehouse_recipe/warehouse_recipes_item.dart
@@ -1,13 +1,18 @@
 import 'package:bakingup_frontend/constants/colors.dart';
 import 'package:bakingup_frontend/models/warehouse.dart';
 import 'package:flutter/material.dart';
+import 'package:shimmer/shimmer.dart';
 
 class WarehouseRecipesItem extends StatelessWidget {
   final List<RecipeItemData> recipeList;
   final int index;
+  final bool isLoading;
 
   const WarehouseRecipesItem(
-      {super.key, required this.recipeList, required this.index});
+      {super.key,
+      required this.recipeList,
+      required this.index,
+      required this.isLoading});
 
   @override
   Widget build(BuildContext context) {
@@ -29,14 +34,30 @@ class WarehouseRecipesItem extends StatelessWidget {
         ),
         child: Row(
           children: [
-            ClipRRect(
-              borderRadius: BorderRadius.circular(13),
-              child: Image.network(
-                recipeList[index].recipeImg,
-                width: 90,
-                height: 60,
-                fit: BoxFit.cover,
-              ),
+            Stack(
+              children: [
+                Shimmer.fromColors(
+                  baseColor: greyColor,
+                  highlightColor: whiteColor,
+                  child: Container(
+                    width: 90,
+                    height: 60,
+                    decoration: BoxDecoration(
+                      color: Colors.white,
+                      borderRadius: BorderRadius.circular(13),
+                    ),
+                  ),
+                ),
+                ClipRRect(
+                  borderRadius: BorderRadius.circular(13),
+                  child: Image.network(
+                    recipeList[index].recipeImg,
+                    width: 90,
+                    height: 60,
+                    fit: BoxFit.cover,
+                  ),
+                ),
+              ],
             ),
             const SizedBox(
               width: 10,
